fix(#73): always probe `which bun`, centralize bun discovery (sage r1)
Sage round 1: 2 important + 1 suggestion. Findings concentrate on
the same defect — the installer pre-flight had a `process.versions.bun`
fast-path bypass.

## Important — pre-flight skipped PATH validation under Bun

`if (process.versions.bun) return;` proved the INSTALLER runs under
Bun. It said nothing about whether `#!/usr/bin/env bun` will resolve
in the hook's later spawn environment (different shell, restricted
PATH, sandboxed substrate). The bypass could pass pre-flight and
still install hooks that fail at fire time.

Replaced with: always `which bun`. `SOMA_BUN_PATH` env override
still wins for unusual deployments. `process.versions.bun` is only
considered as a last-resort fallback in `resolveBunExecutable`
(adapter side), not the pre-flight.

## Suggestion — dedup probe logic

`requireBunInPath` (install.ts) and `resolveBunExecutable` (codex
adapter) had near-identical spawnSync invocations. Extracted to new
`src/bun-probe.ts`:
- `locateBunOnPath()` — shared `which bun` wrapper
- `requireBunInPath()` — installer pre-flight
- `resolveBunExecutable()` — adapter config embed
Future PATH/diagnostics changes update one place.

419 pass. Typecheck + lint clean. No behavior change for users who
already had bun in PATH.
